+++ R/get_LD.R
@@ -42,44 +42,54 @@
 #' @inheritParams echotabix::query_vcf
 #' @inheritParams downloadR::downloader
 #'
-#' @return A symmetric LD matrix of pairwise SNP correlations.
+#' @returns A named list containing:
+#' \itemize{
+#' \item{"LD": }{Symmetric LD matrix of pairwise SNP correlations.}
+#' \item{"DT": }{Standardised query data filtered to only the 
+#' SNPs included in both \code{query_dat} and the LD matrix.}
+#' \item{"RDS_path": }{The path to where the LD matrix was saved.}
+#' } 
 #'
 #' @family LD
-#' @examples
-#' BST1 <- echodata::BST1 
-#' locus_dir <- file.path(tempdir(), echodata::locus_dir)
-#' BST1 <- BST1[seq(1, 50), ] 
-#' LD_list <- echoLD::load_or_create(
-#'     locus_dir = locus_dir,
-#'     query_dat = BST1,
-#'     LD_reference = "1KGphase1"
-#' ) 
 #' @export
-load_or_create <- function(locus_dir,
-                           query_dat,
-                           force_new_LD = FALSE,
-                           LD_reference = c("1KGphase1", "1KGphase3", "UKB"),
-                           target_genome = "hg19",
-                           samples = NULL,
-                           superpopulation = NULL,
-                           local_storage = NULL,
-                           leadSNP_LD_block = FALSE,
-                           fillNA = 0,
-                           verbose = TRUE,
-                           remove_tmps = TRUE,
-                           as_sparse = TRUE,
-                           download_method = "axel",
-                           conda_env = "echoR",
-                           nThread = 1) {
+#' @importFrom echodata mungesumstats_to_echolocatoR
+#' @examples
+#' query_dat <- echodata::BST1[seq(1, 50), ] 
+#' locus_dir <- file.path(tempdir(), echodata::locus_dir) 
+#'
+#' LD_list <- echoLD::get_LD(
+#'     locus_dir = locus_dir,
+#'     query_dat = query_dat,
+#'     LD_reference = "1KGphase1")
+get_LD <- function(query_dat,
+                   locus_dir=tempdir(),
+                   force_new_LD = FALSE,
+                   LD_reference = c("1KGphase1", "1KGphase3", "UKB"),
+                   target_genome = "hg19",
+                   samples = NULL,
+                   superpopulation = NULL,
+                   local_storage = NULL,
+                   leadSNP_LD_block = FALSE,
+                   fillNA = 0,
+                   verbose = TRUE,
+                   remove_tmps = TRUE,
+                   as_sparse = TRUE,
+                   download_method = "axel",
+                   conda_env = "echoR",
+                   nThread = 1) {
+    
     LD_reference <- LD_reference[1]
     RDS_path <- get_rds_path(
         locus_dir = locus_dir,
         LD_reference = LD_reference
     )
-
+    #### Standardise colnames ####
+    query_dat <- echodata::mungesumstats_to_echolocatoR(
+        dat = query_dat,
+        standardise_colnames = TRUE)
     if (file.exists(RDS_path) & (!force_new_LD)) {
         #### Import existing LD ####
-        messager("+  LD:: Previously computed LD_matrix detected.",
+        messager("Previously computed LD_matrix detected.",
             "Importing...", RDS_path,
             v = verbose
         )
@@ -94,7 +104,7 @@ load_or_create <- function(locus_dir,
         )
     } else if (tolower(LD_reference) == "ukb") {
         #### UK Biobank ####
-        LD_list <- LD_ukbiobank(
+        LD_list <- get_LD_UKB(
             query_dat = query_dat,
             locus_dir = locus_dir,
             force_new_LD = force_new_LD,
@@ -109,7 +119,7 @@ load_or_create <- function(locus_dir,
         )
     } else if (tolower(LD_reference) %in% c("1kgphase1", "1kgphase3")) {
         #### 1000 Genomes ####
-        LD_list <- LD_1KG(
+        LD_list <- get_LD_1KG(
             locus_dir = locus_dir,
             query_dat = query_dat,
             local_storage = local_storage,
@@ -126,7 +136,7 @@ load_or_create <- function(locus_dir,
         c(".vcf", ".vcf.gz", ".vcf.bgz")
     ))) {
         #### Custom vcf ####
-        LD_list <- LD_custom(
+        LD_list <- get_LD_custom(
             locus_dir = locus_dir,
             query_dat = query_dat,
             local_storage = local_storage,
@@ -140,7 +150,7 @@ load_or_create <- function(locus_dir,
         )
     } else {
         msg <- paste0(
-            "echoLD:: LD_reference input not recognized.", 
+            "LD_reference input not recognized.", 
             " Must both one of:\n",
             paste0(" - ",
                 c(
@@ -154,4 +164,9 @@ load_or_create <- function(locus_dir,
         stop(msg)
     }
     return(LD_list)
+}
+
+load_or_create <- function(...){
+    .Deprecated("get_LD")
+    get_LD(...)
 }