+++ R/get_LD_1KG.R
@@ -11,53 +11,39 @@
 #'
 #' @param fillNA When pairwise LD (r) between two SNPs is \code{NA},
 #' replace with 0.
-#' @inheritParams load_or_create
-#' @inheritParams echotabix::query_vcf
+#' @param as_sparse Save/return LD matrix as a sparse matrix. 
+#' @inheritParams get_LD 
 #'
-#' @examples
-#' \dontrun{
-#' query_dat <-  echodata::BST1[seq(1, 50), ]
-#' locus_dir <- echodata::locus_dir
-#' locus_dir <- file.path(tempdir(), locus_dir) 
-#' LD_reference <- system.file("extdata", "BST1.1KGphase3.vcf.bgz",
-#'     package = "echoLD"
-#' )
-#' LD_list <- LD_custom(
-#'     locus_dir = locus_dir,
-#'     query_dat = query_dat
-#'     LD_reference = LD_reference
-#' )
-#' }
 #' @family LD
 #' @keywords internal
-#' @importFrom echotabix query_vcf
-LD_custom <- function(locus_dir = tempdir(),
-                      query_dat,
-                      LD_reference,
-                      target_genome = "GRCh37",
-                      superpopulation = NULL,
-                      samples = NULL,
-                      local_storage = NULL,
-                      leadSNP_LD_block = FALSE,
-                      force_new = FALSE,
-                      force_new_MAF = FALSE,
-                      fillNA = 0,
-                      stats = "R",
-                      as_sparse = TRUE,
-                      # min_r2=FALSE,
-                      # min_Dprime=FALSE,
-                      # remove_correlates = FALSE,
-                      verbose = TRUE) {
+get_LD_1KG <- function(locus_dir,
+                   query_dat,
+                   LD_reference = "1KGphase1",
+                   superpopulation = NULL,
+                   samples = NULL,
+                   local_storage = NULL,
+                   leadSNP_LD_block = FALSE,
+                   force_new = FALSE,
+                   force_new_MAF = FALSE,
+                   fillNA = 0,
+                   stats = "R",
+                   as_sparse = TRUE,
+                   # min_r2=FALSE,
+                   # min_Dprime=FALSE,
+                   # remove_correlates = FALSE,
+                   verbose = TRUE) {
     
-    messager("echoLD:: Using custom VCF as LD reference panel.", v = verbose) 
-    target_path <- LD_reference
+    messager("Using 1000Genomes as LD reference panel.", v = verbose)
     #### Query ####
-    vcf <- echotabix::query_vcf(
-        query_granges = query_dat,
-        target_path = target_path, 
-        target_genome = target_genome,
+    query_granges <- echotabix::construct_query(query_dat=query_dat)
+    vcf <- get_LD_1KG_download_vcf(
+        query_granges = query_granges,
+        locus_dir = locus_dir,
+        LD_reference = LD_reference,
+        superpopulation = superpopulation,
         samples = samples,
         force_new = force_new,
+        local_storage = local_storage,
         verbose = verbose
     )
     #### Convert to snpStats object ####
@@ -97,9 +83,9 @@ LD_custom <- function(locus_dir = tempdir(),
     #### Convert to sparse ####
     if(as_sparse){
         LD_matrix <- to_sparse(X = LD_matrix,
-                          verbose = verbose)
+                               verbose = verbose)
     }
-    #### Save LD matrix ####
+    #### Save LD matrix #### 
     LD_list <- save_LD_matrix(
         LD_matrix = LD_matrix,
         dat = query_dat,
@@ -107,6 +93,7 @@ LD_custom <- function(locus_dir = tempdir(),
         subset_common = TRUE,
         fillNA = fillNA,
         LD_reference = LD_reference,
+        as_sparse = as_sparse,
         verbose = verbose
     )
     return(LD_list)