+++ R/get_lead_r2.R
@@ -3,7 +3,7 @@
 #' Add new columns r and r2 containing the degree of LD in each SNP (row)
 #'  with the lead GWAS/QTL SNP. 
 #'
-#' @param query_dat SNP-level data.
+#' @param dat SNP-level data.
 #' @param LD_matrix LD matrix.
 #' @param fillNA Value to fill NAs with in r/r2 columns.
 #' @param LD_format The format of the provided \code{LD_matrix}:
@@ -17,9 +17,9 @@
 #' @importFrom data.table as.data.table merge.data.table
 #' 
 #' @examples
-#' query_dat2 <- echoLD::get_lead_r2(query_dat = echodata::BST1,
-#'                                   LD_matrix = echodata::BST1_LD_matrix)
-get_lead_r2 <- function(query_dat,
+#' dat2 <- echoLD::get_lead_r2(dat = echodata::BST1,
+#'                             LD_matrix = echodata::BST1_LD_matrix)
+get_lead_r2 <- function(dat,
                         LD_matrix = NULL,
                         fillNA = 0,
                         LD_format = "matrix",
@@ -27,10 +27,10 @@ get_lead_r2 <- function(query_dat,
     # Avoid confusing checks
     r <- leadSNP <- NULL;
 
-    if (any(c("r", "r2") %in% colnames(query_dat))) {
-       query_dat<- query_dat[, -c("r", "r2")]
+    if (any(c("r", "r2") %in% colnames(dat))) {
+       dat<- dat[, -c("r", "r2")]
     }
-    LD_SNP <- unique(subset(query_dat, leadSNP)$SNP)
+    LD_SNP <- unique(subset(dat, leadSNP)$SNP)
     if (length(LD_SNP) > 1) {
         LD_SNP <- LD_SNP[1]
         warning("More than one lead SNP found. Using only the first one:",
@@ -50,11 +50,11 @@ get_lead_r2 <- function(query_dat,
 
     if (LD_format == "matrix") {
         if (is.null(LD_matrix)) {
-            messager("+ echoLD:: No LD_matrix detected. Setting r2=NA",
+            messager("No LD_matrix detected. Setting r2=NA",
                      v = verbose) 
-           query_dat$r2 <- NA
+           dat$r2 <- NA
         } else {
-            messager("+ echoLD:: LD_matrix detected.",
+            messager("LD_matrix detected.",
                 "Coloring SNPs by LD with lead SNP.",
                 v = verbose
             )
@@ -65,7 +65,7 @@ get_lead_r2 <- function(query_dat,
                 `colnames<-`(c("SNP", "r")) %>%
                 dplyr::mutate(r2 = r^2) %>%
                 data.table::as.data.table()
-           query_dat<- data.table::merge.data.table(query_dat, LD_sub,
+           dat<- data.table::merge.data.table(dat, LD_sub,
                 by = "SNP",
                 all.x = TRUE
             )
@@ -76,16 +76,16 @@ get_lead_r2 <- function(query_dat,
             `colnames<-`(c("SNP", "r")) %>%
             dplyr::mutate(r2 = r^2) %>%
             data.table::as.data.table()
-       query_dat<- data.table::merge.data.table(query_dat, LD_sub,
+       dat<- data.table::merge.data.table(dat, LD_sub,
             by = "SNP",
             all.x = TRUE
         )
     }
 
-    if ((fillNA != FALSE) & (sum(is.na(LD_matrix)) > 0)) {
-        messager("+ echoLD:: Filling r/r2 NAs with", fillNA, v = verbose)
-        query_dat$r[is.na(query_dat$r)] <- fillNA
-        query_dat$r2[is.na(query_dat$r2)] <- fillNA
+    if ((!isFALSE(fillNA)) & (sum(is.na(dat$r)) > 0)) {
+        messager("Filling r/r2 NAs with", fillNA, v = verbose)
+        dat$r[is.na(dat$r)] <- fillNA
+        dat$r2[is.na(dat$r2)] <- fillNA
     }
-    return(query_dat)
+    return(dat)
 }