+++ R/saveSparse.R
@@ -13,6 +13,6 @@ saveSparse <- function(LD_matrix,
     # print(object.size(LD_matrix),units="auto")
     # messager("Sparse size:")
     # print(object.size(LD_sparse),units="auto")
-    messager("echoLD:: Saving LD as sparse matrix ==>", LD_path, v = verbose)
+    messager("Saving LD as sparse matrix ==>", LD_path, v = verbose)
     saveRDS(LD_sparse, LD_path)
 }